Modification to SH lid.
Modification to "202Q-OQTO-8port-7x7mm-Lid.step" by decreasing the height to 4 mm instead of 7 mm. This was done in order to enable a mu-metal plate to be place within close proximity to the chip. The inner cavity is now 3 mm in height. The change was implemented in FreeCAD for which the corresponding file has been added into the repository. In the following updates, the CAD design shall be converted into FreeCAD projects in order for users to more easily do modifications.